+++ dessia_common/optimization.py
@@ -31,7 +31,7 @@ class BoundedAttributeValue(DessiaObject):
     :param attribute_name: Name of attribute to bound.
     :param min_value: Minimum value for this attribute.
     :param max_value: Maximum value for this attribute.
-    :param number: Number of values to generate betwwen those bounds.
+    :param number: Number of values to generate between those bounds.
         Only used for sampling. ClassSampler.full_fact method.
     :param name: Name of BoundedAttributeValue.
     """
@@ -91,7 +91,7 @@ class DrivenModelOptimizer(Optimizer):
 
 
 class InstantiatingModelOptimizer(Optimizer):
-    """ Abstract class, to be subclassed by real class. Instantiate a new model at each point request. """
+    """ Abstract class, to be inherited by a real class. Instantiate a new model at each point request. """
 
     def __init__(self, fixed_parameters: List[FixedAttributeValue], optimization_bounds: List[BoundedAttributeValue],
                  name: str = ''):