+++ dessia_common/utils/__init__.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Utils package.
+Package that gathers several helper and utility features.
 """